Last round of cleaning up the vestigial files and comments. Refactored the Chart Instructions component logic slightly for better rendering in all CHart layouts.
