+++ client/src/components/ProTx/components/charts/ChartInstructions.js
@@ -16,16 +16,11 @@ function ChartInstructions({
       return (
         <div>
           <div className="chart-description-label">Description</div>
-          {currentDescription}
+          <div className="chart-description">{currentDescription}</div>
         </div>
       );
     }
-    return (
-      <div>
-        <h3>Description</h3>
-        None.
-      </div>
-    );
+    return <div />;
   };
 
   const currentDescription = getCurrentDescription(
@@ -48,7 +43,7 @@ function ChartInstructions({
           <li key={instruction}>{instruction}</li>
         ))}
       </ul>
-      <div>{currentDescription}</div>
+      {currentDescription}
     </div>
   );
 }
