--- client/src/components/ProTx/components/charts/MaltreatmentObservedFeaturesChart.css
@@ -1,1 +1,0 @@
-/* .maltreatment-observed-features-chart{} */
