+++ client/src/components/ProTx/components/dashboard/Dashboard.js
@@ -2,7 +2,7 @@ import React from 'react';
 import DashboardHeader from './DashboardHeader';
 import DashboardDisplay from './DashboardDisplay';
 import '../Variables.css';
-import './AnalyticsDashboard.css';
+import './Dashboard.css';
 
 // Need to find a way to set this value based on some external configuration value so it changes based on the hosts settings.
 // false for PROD, true for all non-PROD deployments.
@@ -11,10 +11,10 @@ const hostName = "COOK·CHILDREN'S·PRE-PRODUCTION·PORTAL";
 const messageString =
   'Geospatial data analytics dashboard using incomplete demonstration data.';
 
-function AnalyticsDashboard() {
+function Dashboard() {
   if (showHeader) {
     return (
-      <div className="analytics-dashboard-root-header">
+      <div className="dashboard-root-header">
         <DashboardHeader hostName={hostName} messageString={messageString} />
         <DashboardDisplay />
       </div>
@@ -22,10 +22,10 @@ function AnalyticsDashboard() {
   }
 
   return (
-    <div className="analytics-dashboard-root">
+    <div className="dashboard-root">
       <DashboardDisplay />
     </div>
   );
 }
 
-export default AnalyticsDashboard;
+export default Dashboard;
