+++ client/src/components/ProTx/index.js
@@ -1,1 +1,1 @@
-export { default } from './components/dashboard/AnalyticsDashboard';
+export { default } from './components/dashboard/Dashboard';
